move tools to crates/runtime/src/tools/
